chore(mobile): Moves linking initialURL, returnToOnAuthPath, and context switching to useHandleLinking hook and promotes to AuthRootNavigator from HomeNavigator
